+++ src/utils.ts
@@ -1,9 +1,4 @@
-import semver from "semver";
 import { workspace } from "vscode";
-
-export function hasHighEnoughReScriptRelayVersion(version: string): boolean {
-  return semver.satisfies(version.replace(/[\^\~]/g, ""), ">=0.13.0");
-}
 
 export function getPreferredFragmentPropName(onType: string): string {
   const result =