+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Code action for adding `@connection`.
 - Look up GraphQL between files to be able to provide (some) hover info when a type isn't accessed in its original file.
+- Enable new experimental features for queries, mutations and subscriptions.
 
 ## 0.6.0